i16 -> 18 or u8 (#29)
@[default:'`反引号包裹`']
```
0	无效 / 禁止 / 停用 / 不可用
1	有效 / 允许 / 启用 / 可用
```

+++ backend/model/mixins/common.v
@@ -4,10 +4,10 @@ import time
 
 pub struct Common {
 pub:
-	updater_id ?string    @[comment: '修改者ID'; omitempty; sql_type: 'CHAR(36)']
-	updated_at time.Time  @[comment: 'Update Time | 修改日期'; omitempty; sql_type: 'TIMESTAMP']
-	creator_id ?string    @[comment: '创建者ID'; immutable; omitempty; sql_type: 'CHAR(36)']
-	created_at time.Time  @[comment: 'Create Time | 创建日期'; immutable; omitempty; sql_type: 'TIMESTAMP']
-	del_flag   i16        @[comment: '删除标记，-1：已删除，0：未删除'; default: 0; omitempty; sql_type: 'smallint']
+	updater_id ?string @[comment: '修改者ID'; omitempty; sql_type: 'CHAR(36)']
+	updated_at time.Time @[comment: 'Update Time | 修改日期'; omitempty; sql_type: 'TIMESTAMP']
+	creator_id ?string @[comment: '创建者ID'; immutable; omitempty; sql_type: 'CHAR(36)']
+	created_at time.Time @[comment: 'Create Time | 创建日期'; immutable; omitempty; sql_type: 'TIMESTAMP']
+	del_flag   i8 @[comment: '删除标记，-1：已删除，0：未删除'; default: 0; omitempty; sql_type: 'smallint']
 	deleted_at ?time.Time @[comment: 'Delete Time | 删除日期'; omitempty; sql_type: 'TIMESTAMP']
 }
